+++ test/quantum/eval.py
@@ -74,3 +74,12 @@ def test_pytorch_consistent_eval(c):
             pure_result
             @ pure_result.conjugate(diagrammatic=False))
         assert is_close_smallno(doubled_result, mixed_result.to_tensor())
+
+
+@pytest.mark.parametrize('c', pure_circuits)
+def test_quimb_pure_eval(c):
+    print(c)
+    t = c.to_quimb().contract()
+    t = t.data.transpose(*np.argsort(t.inds))
+
+    assert np.allclose(t, c.eval().array), f"{t} != {c.eval().array}"
